+++ Sources/SQLMacros/SQLMacro.swift
@@ -323,5 +323,6 @@ extension SQLResultMacro: ExtensionMacro {
         SQLFunctionMacro.self,
         SQLCodecMacro.self,
         SQLRowMacro.self,
+        SQLBindingsMacro.self,
     ]
 }